remove refs to old asset info components and minor refactor of models
